refactor(api): restructure API services by removing the old Api class; introduce BaseApi class and separate services for chat, file, and project management
